pyproject.toml
[project]
name = "dish-booking-agent"
version = "0.1.0"
description = "AI Agent for booking rooms via Dish MCP"
requires-python = ">=3.12"
dependencies = [
    "pydantic-ai>=0.0.14",
    "python-dotenv>=1.0.1",
    "httpx>=0.27.0",
    "openai>=1.55.0",
    "mcp>=1.0.0"
]

# mypy configuration
[tool.mypy]
python_executable = ".venv/bin/python"
show_error_codes = true
exclude = ["docs", "tests", "LICENSE"]
strict = true
namespace_packages = true

# black configuration
[tool.black]
line-length = 100
include = '\.pyi?$'
exclude = '''
/(
	\.git
| \.hg
| \.mypy_cache
| \.tox
| \.venv
| _build
| buck-out
| build
)/
'''

# ruff configuration
[tool.ruff]
# use Python 3.10 as the minimum version for autofixing
target-version = "py310"

# Match black. Note that this also checks comment line length, but black does not format comments.
line-length = 100

show-fixes = true

[tool.ruff.lint]
ignore-init-module-imports = true
select = [
    "C4",  # flake8-comprehensions
    "SIM", # flake8-simplify
    "Q",   # flake8-quotes
    "ISC", # flake8-implicit-str-concat
    "F",   # pyflakes
    "D",   # pydocstyle
    "E",   # pycodestyle error
    "W",   # pycodestyle warning
    "N",   # pep8-naming
    "I",   # isort
    "PL",  # pylint rules from categories "Convention", "Error", and "Warning"
    "PLE", # ruff currently implements only a subset of pylint's rules
    "PLW", # pylint warning
    "PLR", # pylint refactor
    "UP",  # pyupgrade
    "C",   # Complexity (mccabe+) & comprehensions
]
ignore = [
    "UP006", # See https://github.com/bokeh/bokeh/issues/13143
    "UP007", # See https://github.com/bokeh/bokeh/pull/13144
]

[tool.ruff.lint.pydocstyle]
# Use Google-style docstrings.
convention = "google"

[tool.ruff.lint.mccabe]
# Flag errors (`C901`) whenever the complexity level exceeds 5.
max-complexity = 5


# typos configuration
[tool.typos.files]
extend-exclude = [".gitignore", "LICENSE", ".*"]

[tool.typos.default.extend-words]
center = "center"

[tool.typos.default]
locale = "en-gb"
